pyproject.toml
[tool.poetry]
name = "mojap-airflow-tools"
version = "0.0.1"
description = "A few wrappers and tools to use airflow on AP"
authors = ["mojap-data-engineers"]

[tool.poetry.dependencies]
python = "^3.7"
poetry_version = ">=0.1.5"
apache-airflow = {extras = ["kubernetes"], version = "^1.10.10"}

[tool.poetry.extras]
airflow = ["apache-airflow"]

[tool.poetry.dev-dependencies]
pytest = "^5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
